@@ -15,62 +15,60 @@
         "hot": "cross-env NODE_ENV=development webpack-dev-server --inline --hot --config=node_modules/laravel-mix/setup/webpack.config.js",
         "prod": "npm run production",
         "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
-        "lightcrawler-a": "lightcrawler -u http://localhost:8000 -c lightcrawler.json",
-        "lightcrawler-m": "lightcrawler -u http://localhost:8000/en/manager -c lightcrawler.json",
         "storybook": "start-storybook"
     },
     "devDependencies": {
-        "@babel/core": "^7.4.3",
-        "@fortawesome/fontawesome-free": "^5.8.0",
-        "@storybook/addon-actions": "^5.0.6",
-        "@storybook/addon-info": "^5.0.6",
-        "@storybook/addon-knobs": "^5.0.6",
-        "@storybook/react": "^5.0.6",
+        "@babel/core": "^7.4.4",
+        "@fortawesome/fontawesome-free": "^5.8.1",
+        "@storybook/addon-actions": "^5.0.11",
+        "@storybook/addon-info": "^5.0.11",
+        "@storybook/addon-knobs": "^5.0.11",
+        "@storybook/react": "^5.0.11",
         "@types/classnames": "^2.2.7",
-        "@types/node": "^11.9.4",
-        "@types/react": "^16.8.4",
-        "@types/react-dom": "^16.8.2",
+        "@types/node": "^12.0.0",
+        "@types/react": "^16.8.15",
+        "@types/react-dom": "^16.8.4",
         "@types/react-intl": "^2.3.17",
         "@types/storybook__react": "^4.0.1",
-        "@typescript-eslint/eslint-plugin": "^1.4.0",
-        "@typescript-eslint/parser": "^1.4.0",
-        "clone-framework": "^2.2.1",
+        "@typescript-eslint/eslint-plugin": "^1.7.0",
+        "@typescript-eslint/parser": "^1.7.0",
+        "clone-framework": "^2.2.6",
         "cross-env": "^5.2.0",
         "cssnano": "^4.1.7",
-        "eslint": "^5.14.1",
+        "eslint": "^5.16.0",
         "eslint-config-airbnb": "^17.1.0",
-        "eslint-config-prettier": "^4.0.0",
-        "eslint-plugin-import": "^2.16.0",
+        "eslint-config-prettier": "^4.2.0",
+        "eslint-plugin-import": "^2.17.2",
         "eslint-plugin-jsx-a11y": "^6.2.1",
-        "eslint-plugin-node": "^8.0.1",
+        "eslint-plugin-node": "^9.0.1",
         "eslint-plugin-prettier": "^3.0.1",
-        "eslint-plugin-promise": "^4.0.1",
+        "eslint-plugin-promise": "^4.1.1",
         "eslint-plugin-react": "^7.12.4",
         "imagemin": "^6.1.0",
         "jquery": "^3.1.0",
         "laravel-mix": "^4.0.15",
-        "lightcrawler": "0.0.9",
-        "node-sass": "^4.11.0",
-        "prettier": "^1.16.4",
+        "node-sass": "^4.12.0",
+        "prettier": "^1.17.0",
         "react-docgen-typescript-loader": "^3.1.0",
-        "resolve-url-loader": "^3.0.1",
+        "resolve-url-loader": "^3.1.0",
         "sass-loader": "^7.1.0",
-        "ts-loader": "^5.3.3",
-        "typescript": "^3.3.3",
-        "vue-template-compiler": "^2.6.7",
-        "webpack": "^4.29.5"
+        "ts-loader": "^6.0.0",
+        "typescript": "^3.4.5",
+        "vue-template-compiler": "^2.6.10",
+        "webpack": "^4.30.0"
     },
     "dependencies": {
         "axios": "^0.18.0",
         "classnames": "^2.2.6",
+        "core-js": "^3.0.1",
         "es6-promise": "^4.2.5",
         "lodash": "^4.17.11",
         "moment": "^2.24.0",
-        "react": "^16.8.3",
-        "react-app-polyfill": "^0.2.2",
-        "react-dom": "^16.8.4",
+        "react": "^16.8.6",
+        "react-app-polyfill": "^1.0.0",
+        "react-dom": "^16.8.6",
         "react-intl": "^2.8.0",
-        "sweetalert2": "^8.2.2"
+        "sweetalert2": "^8.9.0"
     },
     "repository": {
         "type": "git",
