@@ -16,7 +16,7 @@ class FullFrameGuard
 
         $response = $next($request);
 
-        $response->headers->set('X-Frame-Options', 'DENY');
+        $response->headers->set('X-Frame-Options', 'SAMEORIGIN');
 
         return $response;
     }
