@@ -7,7 +7,9 @@
 //require('./components/JobPoster');
 //require('./components/Reference');
 //require('./components/ReferenceList');
-require("./components/ApplicationReview/ReviewApplicationsContainer");
-require("./components/ApplicationReview/ApplicationReviewContainer");
+require('./components/ApplicationReview/ReviewApplicationsContainer');
+require('./components/ApplicationReview/ApplicationReviewContainer');
+require('./components/ApplicantSkills/WordCounter');
+require('./components/ApplicantSkills/SkillsInfoModal');
 
 //import ReviewApplications from './components/ReviewApplications';
