@@ -85,10 +85,12 @@ return [
     "claim_title" => "Basic Information",
     "select_default_option" => "Select from the following...",
     "language_declaration" => "Do you meet the language requirements for this position?",
-    "english_essential" => "I understand that this position has an English essential language requirement. This means that the manager may choose to test my proficiency in English.",
-    "french_essential" => "I understand that this position has a French essential language requirement. This means that the manager may choose to test my proficiency in French.",
-    "bilingual" => 'I understand that this position has a bilingual language requirement. Knowledge of both English and French are required. I am entitled to be assessed on the "Need to Have" and "Nice to Have" criteria for this position in the Official Language of my choice (English or French), however the bilingual requirement means I will have to do additional testing in my second language to confirm my language ability.',
-    "english_french" => "COPY NEEDED This should not be seen",
+    "english_essential" => "I have strong reading, writing and verbal communication skills in English.",
+    "french_essential" => "I have strong reading, writing and verbal communication skills in French.",
+    "english_or_french" => 'I have strong reading, writing and verbal communication skills in English or French.',
+    "bilingual_intermediate" => 'I have intermediate or better reading, writing and verbal communication skills in English and French.',
+    "bilingual_advanced" => 'I have strong reading, writing and verbal communication skills in English and French.',
+    "language_declaration_error_message" => "This box must be checked to proceed.",
     "language_title" => "Language Selection",
     "language_copy" => "Which language would you prefer for this application process?",
     "language_label" => "Select One",
@@ -128,14 +130,19 @@ return [
         "add_button_label" => "Add Skill",
         "null_copy" => "You don't currently have any skills on your profile! Use the button above to add a skill."
     ],
-    "essential_title" => "Skills You Need to Have",
+    /* Step 3: Essential SKills Header */
+    "essential_title" => "Need to Have",
+    "essential_subtitle" => "Remember, that you need ALL these skills to apply for this job!",
+    "essential_question_1" => "Q. Really? I have most of these skills and I’m a quick learner.",
+    "essential_answer_1" => "Sorry, you really do need all of them, and at the right level.",
+    "essential_question_2" => "Q. What if I stretch the truth a little? Once I’m interviewed they’ll change their mind for sure!",
+    "essential_answer_2" => "Don’t do it. These will be assessed, and you’re better than that.",
+    "essential_sidebar_label" => "Skills Checklist",
+
     "asset_title" => "Skills That Are Nice to Have",
-    "essential_context" => "The job you’re about to apply to requires the following skills. Yes, really, all of them. Each one at the minimum level of expertise specified by the hiring manager. Even if you’re amazing and you’re only missing one of these skills, you won’t be considered for this job.  And you have to provide evidence against any skill claims you make… claims that become part of a government record. So don’t be tempted to stretch the truth just to apply. You’re better than that. Save yourself the time and consider applying for something else instead.",
     "asset_context" => "You can apply for this job if you don’t have any of these skills. These are assets that the hiring manager thinks will help with the work. They’re often used to help select the best candidate amongst all those who meet all the “Need to Have” criteria. So if you do have these skills, now is the time to let the hiring manager know.",
-    "essential_start_button_title" => "Scroll to begin filling out the skills you need to have.",
     "asset_start_button_title" => "Scroll to begin filling out the skills that are nice to have.",
     "skills_start_button_label" => "Get Started",
-    "essential_sidebar_label" => "Skills Checklist",
     "asset_sidebar_label" => "Skills Checklist",
     /* Step 05 */
     "step_05_title" => "Talent Cloud | Applicant: Apply Now - Step 5",
