@@ -32,7 +32,7 @@ return [
     /* <title> tag content */
     "home_title" => "Talent Cloud | Manager: Home",
     "profile_title" => "Talent Cloud | Manager: My Profile",
-    "job_index_title" => "Talent Cloud | Manager: My Job Posts",
+    "job_index_title" => "Talent Cloud | Manager: My Job Posters",
     "job_create_title" => "Talent Cloud | Manager: Create a Job Poster",
     # Button labels
     "mobile_trigger_label" => "Mobile Menu",
