@@ -9,26 +9,26 @@ return [
     * The following language lines are used in the applicant index.
     *
     */
-    "title" => "Mes demandes d'emploi",
-    "application_index_title" => "Nuage de talents | Candidat : Mes demandes d'emploi",
-    "progress_title" => "Demandes d'emploi en cours",
-    "completed_title" => "Demandes d'emploi soumises",
-    "expired_title" => "Demandes d'emploi expirées",
+    "title" => "Mes applications",
+    "application_index_title" => "Nuage de talents | Candidat : Mes applications",
+    "progress_title" => "Applications en cours",
+    "completed_title" => "Applications soumises",
+    "expired_title" => "Applications expirées",
     "apply_by_label" => "Postuler avant le",
     "days_remaining_label" => ":count jours restant",
     "time_remaining" => ":time restant",
     "job_closed" => "L'affiche est fermée",
     "edit_date_label" => "Dernière modification",
     "post_label" => "Affiche d'emploi",
-    "post_link_title" => "Voir l'affiche pour cette demandes d'emploi",
+    "post_link_title" => "Voir l'affiche pour cette applications",
     "post_link_label" => "Voir",
     "manager_profile_link_title" => "Voir le profil de ce superviseur.",
     "manager_profile_link_label" => "Voir profil",
-    "draft_link_title" => "Modifier cette demande d'emploi.",
-    "draft_link_label" => "Modifier cette demande d'emploi.",
-    "view_link_title" => "Voir la demande d'emploi.",
-    "view_link_label" => "Voir la demande d'emploi.",
-    "delete_title" => "Supprimer cette demande d'emploi.",
+    "draft_link_title" => "Modifier cette application.",
+    "draft_link_label" => "Modifier cette application.",
+    "view_link_title" => "Voir l'application.",
+    "view_link_label" => "Voir l'application.",
+    "delete_title" => "Supprimer cette application.",
     "delete_label" => "Supprimer",
     "null" => [
         "title" => "Il y a aucune applications."
@@ -36,9 +36,9 @@ return [
     "modals" => [
         "delete_application" => [
             "type" => "deleteConfirmation",
-            "title" => "Supprimer la demande d'emploi.",
+            "title" => "Supprimer l'application.",
             "content" => [
-                "00" => "Êtes-vous sûr de vouloir supprimer cette demande d'emploi? Tous votre progrès sera perdus.",
+                "00" => "Êtes-vous sûr de vouloir supprimer cette application? Tous votre progrès sera perdus.",
                 "01" =>  "Les expériences et compétences sauvegardées resteront accessibles sur votre profil pour une utilisation ultérieure."
             ],
             "id" => "deleteTrash",
