@@ -2,13 +2,13 @@
 
 return [
     /*
-     * --------------------------------------------------------------------------
-     * FAQ Language Lines
-     * --------------------------------------------------------------------------
-     *
-     * The following language lines are used on the FAQ page.
-     *
-     */
+    * --------------------------------------------------------------------------
+    * FAQ Language Lines
+    * --------------------------------------------------------------------------
+    *
+    * The following language lines are used on the FAQ page.
+    *
+    */
     "title" => "Nuage de talents : mode d'emploi",
     "faq_title" => "Nuage de talents | Candidat : FAQ",
     "sidebar_label" => "Sur cette page",
@@ -18,37 +18,37 @@ return [
         "category_title" => "Choses à savoir",
         "sections" => [
             "00" => [
-                "title" => "Pourquoi le Canada a-t-il besoin d’une plateforme pour le travail axé sur les projets?",
+                "title" => "Pourquoi le Canada a-t-il besoin d'une plateforme pour le travail axé sur les projets?",
                 "copy" => [
-                    "00" => "Au XXe siècle, les travailleurs se classaient en très grande majorité dans deux catégories : les travailleurs permanents (qui « appartenaient » à l’organisation, mais qui avaient des avantages sociaux); ou à contrat (travailleurs autonomes sans avantages sociaux). Avec la montée d’Internet, un type d’emploi différent est cependant apparu pour suivre le rythme des nouvelles économies débloquées par l’ère numérique. Parfois appelé « emploi à la demande », emploi axé sur des projets ou emploi non traditionnel, cette nouvelle catégorie de travail représente maintenant près de 30 % de l’emploi au Canada et aux États-Unis, avec des augmentations semblables en Europe. Cependant, malgré la croissance fulgurante du travail axé sur les projets, les gouvernements n’ont pas adapté leurs modèles d’embauche.",
-                    "01" => "Pour sortir des vieilles dualités, le Nuage de talents met à l’essai l’introduction d’une troisième catégorie de travailleurs, à savoir un employé à la demande ayant des droits et des avantages, sur un cheminement de carrière autodirigé et à mobilité élevée. Pour que cela fonctionne, les candidats ont besoin d’une plateforme qui leur permet de trouver des possibilités et d’être valorisés de différentes façons, et les gestionnaires ont besoin d’un modèle de dotation optimisé pour l’ère numérique. Bien que le gouvernement du Canada ait des travailleurs « temporaires » depuis longtemps, il s’agit du premier modèle de talent du secteur public créé, de la technologie à la conception comportementale, pour permettre un marché florissant pour le travail axé sur des projets."
+                    "00" => "Au XXe siècle, la plupart des travailleurs se classaient dans deux catégories : les travailleurs permanents (qui « appartenaient » à l'organisation, mais qui avaient des avantages sociaux); ou à contrat (travailleurs autonomes sans avantages sociaux). Internet a débloqué de nouvelles économies, ainsi qu'une augmentation de l'emploi axé sur des projets ou de l'emploi non traditionnel. Parfois appelé « emploi à la demande », cette nouvelle catégorie de travail représente maintenant près de 30 % de l'emploi au Canada et aux États-Unis, avec des tendances semblables en Europe. Malgré ce changement radical, les gouvernements n'ont pas adapté leurs modèles d'embauche.",
+                    "01" => "Pour sortir des vieilles dualités, le Nuage de talents met à l'essai l'introduction d'une troisième catégorie de travailleurs, à savoir un employé à la demande, sur un cheminement de carrière autodirigé et à mobilité élevée, mais avec les droits et avantages qui ont traditionnellement été réservés aux employés permanents. Pour que cela fonctionne, les demandeurs ont besoin d'une plateforme qui leur permet de trouver des possibilités et de démontrer facilement leur valeur, et les gestionnaires ont besoin d'un modèle de dotation optimisé pour l'ère numérique. Bien que le gouvernement du Canada ait des travailleurs « temporaires » depuis longtemps, il s'agit du premier modèle de talent du secteur public créé pour permettre un marché florissant pour le travail axé sur des projets."
                 ]
             ],
             "01" => [
                 "title" => "Qui peut présenter une demande?",
                 "copy" => [
-                    "00" => "N’importe qui dans le monde peut postuler au moyen du Nuage de talents. Tous les emplois à cette étape du projet sont ouverts au public et publiés de façon concurrentielle.",
+                    "00" => "N'importe qui dans le monde peut postuler au moyen du Nuage de talents. Tous les emplois à cette étape du projet sont ouverts au public et publiés de façon concurrentielle.",
                     "01" => "En vertu des politiques de dotation du gouvernement du Canada, les anciens combattants canadiens et les citoyens canadiens auront la préférence en matière de dotation (dans le jargon non gouvernemental, cela signifie que si vous êtes résident(e) permanent(e) ou non canadien(ne), votre demande sera prise en considération si aucun citoyen canadien qualifié ne pose sa candidature au poste… et il y aura ensuite des formalités supplémentaires)."
                 ]
             ],
             "02" => [
-                "title" => "Quels types d’emplois seront publiés?",
+                "title" => "Quels types d'emplois seront publiés?",
                 "copy" => [
-                    "00" => "Tous les postes publiés dans le Nuage de talents pour le projet-pilote sont des postes pour une période déterminée. Cela signifie que vous êtes un(e) employé(e) du gouvernement du Canada pour la durée de votre mandat et que vous pouvez accumuler des prestations après trois mois et une pension après six mois d’emploi continu. Il est possible de prendre des mandats séquentiels et de les jumeler à un emploi continu au sein du gouvernement du Canada, tout en entreprenant différents projets dans différents ministères, si vous le souhaitez. Les postes à durée déterminée peuvent être prolongés jusqu’à trois ans. Après trois ans, les employés nommés pour une période déterminée deviennent des employés nommés pour une période indéterminée (ou permanente), sauf dans des circonstances exceptionnelles.",
-                    "01" => "Le Nuage de talents commence par offrir des emplois liés au numérique, à la technologie et à l’expérience de l’utilisateur. Cherchez donc des postes comme « scientifique des données », « programmeur », « gestionnaire de projet technique », « concepteur de l’expérience utilisateur », etc.",
-                    "02" => "Si la plateforme fonctionne bien (rappelez-vous : il s’agit d’une grande expérience), nous lancerons des volets de dotation dans les postes administratifs et les ressources humaines (RH) vers janvier. Les postes de gestion des politiques et des projets seront lancés vers le mois de mars."
+                    "00" => "Tous les emplois publiés dans le Nuage de talents pendant le projet-pilote sont des postes pour une période déterminée. Cela signifie que vous êtes un(e) employé(e) du gouvernement du Canada pour la durée de votre contrat et que vous pouvez commencer à accumuler des prestations après trois mois et une pension après six mois d'emploi continu. Il est possible de prendre des mandats séquentiels et de les jumeler à un emploi continu au sein du gouvernement du Canada, tout en entreprenant différents projets dans différents ministères, si vous le souhaitez. Les postes à durée déterminée peuvent être prolongés jusqu'à trois ans. Après trois ans au même poste, les employés nommés pour une période déterminée deviennent des employés nommés pour une période indéterminée (permanente), sauf dans des circonstances exceptionnelles.",
+                    "01" => "Le Nuage de talents a commencé par offrir des emplois liés au numérique, à la technologie, et à l'expérience de l'utilisateur. Pensez à des postes comme « scientifique des données », « programmeur », « gestionnaire de projet technique », et « concepteur de l'expérience utilisateur ».",
+                    "02" => "Depuis lors, nous avons lancé des volets de dotation dans les postes administratifs et les ressources humaines (RH). Ensuite, nous élargirons notre champ d'action pour y inclure les postes en matière de politiques, de gestion de projets et d'approvisionnement, que nous prévoyons publier plus tard ce printemps."
                 ]
             ],
             "03" => [
-                "title" => "Pourquoi y a-t-il si peu d’emplois publiés dans le Nuage de talents?",
+                "title" => "Pourquoi y a-t-il si peu d'emplois publiés dans le Nuage de talents?",
                 "copy" => [
-                    "00" => "En tout temps, le Nuage de talents devrait publier de deux à trois postes. Il est également possible que le site doive fermer le lundi matin de temps à autre. Nous sommes une petite équipe qui travaille à un budget restreint sur une idée hautement expérimentale. La publication de quelques emplois à la fois est donc la meilleure façon pour nous de recueillir des données de qualité pour savoir si la plateforme fonctionne comme il se doit, tout en offrant une expérience de qualité (nous l’espérons) aux candidats, aux gestionnaires d’embauche et aux conseillers ministériels en RH."
+                    "00" => "En tout temps, le Nuage de talents devrait publier de deux à trois postes. Il est également possible que le site doive fermer le lundi matin de temps à autre. Nous sommes une petite équipe qui travaille à un budget restreint sur une idée hautement expérimentale. La publication de quelques emplois à la fois est donc la meilleure façon pour nous de recueillir des données de qualité pour savoir si la plateforme fonctionne comme il se doit, tout en offrant ce que nous espérons être une expérience de qualité aux candidats, aux gestionnaires d'embauche et aux conseillers ministériels en RH."
                 ]
             ],
             "04" => [
                 "title" => "Quels ministères participent à ce modèle de dotation expérimental?",
                 "copy" => [
-                    "00" => "Comme il s’agit d’une expérience, le Nuage de talents est financé par des protocoles d’entente (PE) avec les ministères partenaires. Il ne s’agit pas de frais de service pour le personnel qui utilise la plateforme. Il s’agit d’un groupe de ministères qui cofinancent une solution de rechange au modèle traditionnel des RH. Seuls les ministères partenaires peuvent utiliser le Nuage de talents dans le cadre de ce modèle de financement.",
+                    "00" => "Le Nuage de talents est financé par des protocoles d'entente (PE) avec les ministères partenaires. Il ne s'agit pas de frais de service pour le personnel qui utilise la plateforme. Il s'agit d'un groupe de ministères qui cofinancent une solution de rechange expérimentale au modèle traditionnel des RH. Selon ce modèle, seuls les ministères partenaires peuvent embaucher du personnel à l'aide du Nuage de talents.",
                     "01" => "Les ministères partenaires qui ont signé des PE sont les suivants :",
                     "02" => [
                         "type" => "ul",
@@ -62,34 +62,35 @@ return [
                             "06" => "Agence des services frontaliers du Canada",
                             "07" => "Pêches et Océans Canada",
                             "08" => "Innovation, Sciences et Développement économique Canada",
-                            "09" => "Services publics et Approvisionnement Canada"
+                            "09" => "Condition féminine Canada",
+                            "10" => "Services publics et Approvisionnement Canada",
+                            "11" => "Services partagés Canada",
+                            "12" => "Ministère de la Défense nationale"
                         ]
                     ]
                 ]
             ],
             "05" => [
-                "title" => "À quoi puis-je m’attendre dans le processus de demande et de sélection?",
+                "title" => "À quoi puis-je m'attendre dans le processus de demande et de sélection?",
                 "copy" => [
-                    "00" => "Ultimement, nous espère que le Nuage de talents pourra offrir une expérience d’embauche de 30 jours (de l’annonce d’emploi à la lettre d’offre). Pour l’expérience initiale, les candidats doivent s’attendre à recevoir des nouvelles au sujet de leur demande initiale dans les deux semaines suivant la fermeture de l’annonce d’emploi. Pour ceux qui passent à d’autres séries de présélection dans un poste qui a été annoncé comme étant bilingue, il faut s’attendre à des tests de langue tôt dans le processus (si vous n’avez jamais passé d’examen linguistique auprès du gouvernement du Canada auparavant, vous voudrez peut-être essayer ces <a href=\"https://www.canada.ca/fr/commission-fonction-publique/services/evaluation-langue-seconde/tests-autoevaluation.html\" target=\"__blank\" rel=\"noopener noreferrer\" title=\"Ouvrir les tests\">tests d’autoévaluation</a> pour avoir une idée de la nature des tests).",
-                    "01" => "Pour ce qui est de la présélection d’autres compétences, vous pouvez également vous attendre à d’autres approches d’évaluation et de mise à l’essai tout au long du processus qui sont plus informelles et conçues pour mettre à l’essai le bon ajustement à l’équipe et la capacité globale. Le Nuage de talents encourage les gestionnaires à appliquer une approche de sélection « optimale dans l’ensemble » plutôt qu’une approche de sélection « à chaque étape »."
+                    "00" => "Le Nuage de talents travaille à l'élaboration d'une expérience d'embauche de 30 jours de l'annonce d'emploi à la lettre d'offre. Nous sommes actuellement en train de mettre à l'essai le projet. Donc, en ce moment, les candidats doivent s'attendre à recevoir des nouvelles au sujet de leur demande initiale dans les deux semaines suivant la fermeture de l'annonce d'emploi. Pour ceux qui passent l'étape de la présélection initiale dans un poste qui a été annoncé comme étant bilingue, il faut s'attendre à des tests de langue tôt dans le processus (si vous n'avez jamais passé d'examen linguistique auprès du gouvernement du Canada auparavant, vous voudrez peut-être essayer ces <a href=\"https://www.canada.ca/fr/commission-fonction-publique/services/evaluation-langue-seconde/tests-autoevaluation.html\" target=\"__blank\" rel=\"noopener noreferrer\" title=\"Ouvrir les tests\">tests d'autoévaluation</a> pour avoir une idée de la nature des tests).",
+                    "01" => "Vous pouvez également vous attendre à d'autres approches d'évaluation et de mise à l'essai tout au long du processus qui sont plus informelles et conçues pour dépister d'autres compétences, mettre à l'essai le bon ajustement à l'équipe et la capacité globale. Le Nuage de talents encourage les gestionnaires à appliquer une approche de sélection « optimale dans l'ensemble » plutôt que de « présélectionner quelques candidats à chaque étape ».",
+                    "02" => "Tous les employés du gouvernement du Canada sont tenus de réussir les vérifications de la fiabilité, et bon nombre d'entre eux devront obtenir une autorisation de niveau secret. Il s'agit de donner des empreintes digitales, ainsi que de faire vérifier son casier judiciaire et son crédit. Ces fonctions sont exercées par des agents de sécurité autorisés au gouvernement (et non par le Nuage de talents).",
+                    "03" => "Si vous avez vécu à l'extérieur du Canada pendant plus de six mois consécutifs au cours des cinq dernières années (dix ans, si votre emploi exige une cote de sécurité de niveau secret), vous devrez être soumis(e) à une vérification du casier judiciaire « à l'extérieur du pays ». Cela peut prendre un certain temps, car cela dépend des délais de traitement du pays où vous avez vécu. Les demandeurs peuvent demander cette vérification du casier judiciaire « à l'extérieur du pays » en tout temps, et la demande n'a pas besoin d'être liée à un processus de dotation. Ce dossier est présenté par le demandeur au gouvernement du Canada au moment du contrôle de sécurité. Donc, si vous avez passé huit mois en Afrique du Sud ou avez été dans les Alpes pendant un an ou à l'école en Australie, nous vous suggérons de demander vos dossiers dès le départ. Cela peut vous faire gagner beaucoup de temps, à vous et à votre gestionnaire d'embauche, plus tard (et si un poste axé sur un projet en particulier commence à une date fixe, le manque d'autorisation de sécurité à temps peut forcer le gestionnaire d'embauche à sélectionner un autre candidat)."
                 ]
             ],
             "06" => [
-                "title" => "Quels postes devrais-je postuler?",
+                "title" => "Comment décider si je dois postuler ou non?",
                 "copy" => [
-                    "00" => "Lisez les sections sur la culture d’équipe, le milieu de travail et le gestionnaire d’embauche. Réfléchissez sérieusement à la question de savoir si vous seriez un bon candidat. Essayez de chercher et de postuler des postes qui vous permettront de donner de votre mieux à l’équipe et au projet chaque jour. Posez votre candidature à des emplois qui vous satisferont, dans un environnement opérationnel qui vous aidera à prospérer."
+                    "00" => "L'idée d'un emploi au gouvernement peut être très attrayante. Cela amène beaucoup de gens à postuler pour beaucoup d'emplois pour lesquels ils ne sont pas qualifiés ou des emplois qu'ils n'accepteraient pas s'ils en savaient davantage sur le poste au départ. Toutes ces candidatures supplémentaires bloquent le système et ralentissent douloureusement la dotation en personnel.",
+                    "01" => "Le Nuage de talents communique beaucoup plus d'information aux candidats au sujet du poste et au gestionnaire d'embauche pour aider les candidats à prendre part à la discussion sur le choix de la bonne personne. Si un poste exige de la collaboration et que vous préférez travailler seul, ce n'est peut-être pas bon pour vous. Si un poste est occupé à un rythme rapide et que vous aimez prendre de votre temps pour accomplir les tâches, envisagez plutôt de poser votre candidature pour un autre poste dans le Nuage de talents.",
+                    "02" => "Lisez les sections sur la culture d'équipe, le milieu de travail et le gestionnaire d'embauche. Réfléchissez sérieusement à la question de savoir si vous seriez un bon candidat. Assurez-vous de chercher et de postuler à des emplois qui vous combleront et dans un environnement opérationnel qui vous aidera à vous épanouir afin que vous puissiez donner le meilleur de vous-même à l'équipe et au projet chaque jour!"
                 ]
             ],
             "07" => [
-                "title" => "Quand devrais-je m’abstenir de présenter une demande?",
+                "title" => "S'agit-il d'un marché pour les nominations à un poste occasionnel?",
                 "copy" => [
-                    "00" => "L’idée d’un emploi au gouvernement peut être très attrayante. Cela amène beaucoup de gens à postuler pour beaucoup d’emplois pour lesquels ils ne sont pas qualifiés ou des emplois qu’ils n’accepteraient pas s’ils en savaient davantage sur le poste au départ. Toutes ces candidatures supplémentaires bloquent le système et ralentissent la dotation à un rythme douloureusement lent. Le Nuage de talents communique beaucoup plus d’information aux candidats au sujet du poste et au gestionnaire d’embauche pour aider les candidats à prendre part à la discussion sur le choix de la bonne personne. Si un poste exige de la collaboration et que vous préférez travailler seul, ce n’est peut-être pas bon pour vous. Si un poste est occupé à un rythme rapide et que vous aimez prendre de votre temps pour accomplir les tâches, envisagez plutôt de poser votre candidature au poste suivant."
-                ]
-            ],
-            "08" => [
-                "title" => "S’agit-il d’un marché pour les nominations à un poste occasionnel?",
-                "copy" => [
-                    "00" => "Au gouvernement, une « nomination occasionnelle » désigne un poste d’une durée de 90 jours qui peut être accordée à un travailleur au mérite sans concours. Ceux qui occupent des postes occasionnels n’ont pas droit à une pension ou à des avantages sociaux et ne peuvent occuper qu’un poste occasionnel par année dans un ministère. Certains travailleurs cumulent les emplois du gouvernement en prenant une série d’emplois occasionnels au cours de l’année. En raison de l’engagement du Nuage de talents à faire avancer les droits et les avantages sociaux des travailleurs, les nominations occasionnelles ne sont pas permises sur la plateforme."
+                    "00" => "Au gouvernement, une « nomination occasionnelle » désigne un poste d'une durée de 90 jours qui peut être accordée à un travailleur au mérite sans concours. Ceux qui occupent des postes occasionnels n'ont pas droit à une pension ou à des avantages sociaux et ne peuvent occuper qu'un poste occasionnel par année dans un ministère. Certains travailleurs cumulent les emplois du gouvernement en prenant une série d'emplois occasionnels au cours de l'année. En raison de l'engagement du Nuage de talents à faire avancer les droits et les avantages sociaux des travailleurs, les nominations occasionnelles ne sont pas permises sur la plateforme."
                 ]
             ]
         ]
@@ -97,159 +98,159 @@ return [
     "speed" => [
         "sidebar_title" => "Après avoir postulé",
         "category_anchor" => "post-application",
-        "category_title" => "Comment accélérer les choses après la présentation d’une candidature",
+        "category_title" => "Comment accélérer les choses après la présentation d'une candidature",
         "sections" => [
             "00" => [
                 "copy" => [
-                    "00" => "Tous les employés du gouvernement du Canada sont tenus de réussir les vérifications de la fiabilité, et bon nombre d’entre eux devront obtenir une autorisation de niveau secret. Il s’agit de donner des empreintes digitales, ainsi que de faire vérifier son casier judiciaire et son crédit. Ces fonctions sont exercées par des agents de sécurité autorisés au gouvernement (et non par le Nuage de talents).",
+                    "00" => "Tous les employés du gouvernement du Canada sont tenus de réussir les vérifications de la fiabilité, et bon nombre d'entre eux devront obtenir une autorisation de niveau secret. Il s'agit de donner des empreintes digitales, ainsi que de faire vérifier son casier judiciaire et son crédit. Ces fonctions sont exercées par des agents de sécurité autorisés au gouvernement (et non par le Nuage de talents).",
                     "01" => "Tout le processus peut prendre du temps.",
-                    "02" => "Voulez-vous l’accélérer?",
+                    "02" => "Voulez-vous l'accélérer?",
                     "03" => "Il y a deux choses que vous pouvez faire :",
                     "04" => [
                         "type" => "ul",
                         "items" => [
-                            "00" => "Remplissez les documents que vous devrez soumettre à l’avance.",
-                            "01" => "Si vous avez vécu à l’extérieur du Canada, lisez ce qui suit."
+                            "00" => "Remplissez les documents que vous devrez soumettre à l'avance.",
+                            "01" => "Si vous avez vécu à l'extérieur du Canada, lisez ce qui suit."
                         ]
                     ],
-                    "05" => "Si l’emploi pour lequel vous postulez exige une cote de fiabilité, remplissez le <a href=\"http://www.tbs-sct.gc.ca/tbsf-fsct/330-23-fra.asp\" target=\"__blank\" rel=\"noopener noreferrer\" title=\"Ouvrez le formulaire.\">formulaire qui se trouve ici</a> et conservez-le. Lorsque vous aurez terminé, ne l’envoyez pas par courriel au Nuage de talents jusqu’à ce qu’on vous le demande.",
-                    "06" => "Si l’emploi pour lequel vous postulez exige une cote de sécurité de niveau Secret, vous devrez remplir un <a href=\"http://www.tbs-sct.gc.ca/tbsf-fsct/330-23-fra.asp\" target=\"__blank\" rel=\"noopener noreferrer\" title=\"Ouvrez le formulaire.\">formulaire de vérification de la fiabilité</a> et un <a href=\"http://www.tbs-sct.gc.ca/tbsf-fsct/330-60-fra.asp\" target=\"__blank\" rel=\"noopener noreferrer\" title=\"Ouvrez le formulaire.\">formulaire d’attestation de sécurité de niveau Secret</a>. Une fois ces deux formulaires remplis, ne les envoyez pas par courriel au Nuage de talents avant qu’on vous les demande.",
-                    "07" => "Si vous avez vécu à l’extérieur du Canada pendant plus de six mois consécutifs au cours des cinq dernières années (dix ans, si votre emploi exige une cote de sécurité de niveau secret), vous devrez être soumis(e) à une vérification du casier judiciaire « à l’extérieur du pays ». Cela peut prendre un certain temps, car cela dépend des délais procéduraux du pays où vous avez vécu. Les candidats peuvent demander cette vérification du casier judiciaire « à l’extérieur du pays » en tout temps, en dehors de tout processus de dotation. Ce dossier est présenté par le candidat au gouvernement du Canada au moment du contrôle de sécurité. Donc, si vous avez passé huit mois en Afrique du Sud ou avez été dans les Alpes pendant un an ou à l’école en Australie, nous vous suggérons de demander vos dossiers dès le départ. Cela peut vous faire gagner beaucoup de temps, à vous et à votre gestionnaire d’embauche, plus tard (et si un poste axé sur un projet en particulier commence à une date fixe, le manque d’autorisation de sécurité à temps peut forcer le gestionnaire d’embauche à sélectionner un autre candidat)."
+                    "05" => "Si l'emploi pour lequel vous postulez exige une cote de fiabilité, remplissez le <a href=\"http://www.tbs-sct.gc.ca/tbsf-fsct/330-23-fra.asp\" target=\"__blank\" rel=\"noopener noreferrer\" title=\"Ouvrez le formulaire.\">formulaire qui se trouve ici</a> et conservez-le. Lorsque vous aurez terminé, ne l'envoyez pas par courriel au Nuage de talents jusqu'à ce qu'on vous le demande.",
+                    "06" => "Si l'emploi pour lequel vous postulez exige une cote de sécurité de niveau Secret, vous devrez remplir un <a href=\"http://www.tbs-sct.gc.ca/tbsf-fsct/330-23-fra.asp\" target=\"__blank\" rel=\"noopener noreferrer\" title=\"Ouvrez le formulaire.\">formulaire de vérification de la fiabilité</a> et un <a href=\"http://www.tbs-sct.gc.ca/tbsf-fsct/330-60-fra.asp\" target=\"__blank\" rel=\"noopener noreferrer\" title=\"Ouvrez le formulaire.\">formulaire d'attestation de sécurité de niveau Secret</a>. Une fois ces deux formulaires remplis, ne les envoyez pas par courriel au Nuage de talents avant qu'on vous les demande.",
+                        "07" => "Si vous avez vécu à l'extérieur du Canada pendant plus de six mois consécutifs au cours des cinq dernières années (dix ans, si votre emploi exige une cote de sécurité de niveau secret), vous devrez être soumis(e) à une vérification du casier judiciaire « à l'extérieur du pays ». Cela peut prendre un certain temps, car cela dépend des délais procéduraux du pays où vous avez vécu. Les candidats peuvent demander cette vérification du casier judiciaire « à l'extérieur du pays » en tout temps, en dehors de tout processus de dotation. Ce dossier est présenté par le candidat au gouvernement du Canada au moment du contrôle de sécurité. Donc, si vous avez passé huit mois en Afrique du Sud ou avez été dans les Alpes pendant un an ou à l'école en Australie, nous vous suggérons de demander vos dossiers dès le départ. Cela peut vous faire gagner beaucoup de temps, à vous et à votre gestionnaire d'embauche, plus tard (et si un poste axé sur un projet en particulier commence à une date fixe, le manque d'autorisation de sécurité à temps peut forcer le gestionnaire d'embauche à sélectionner un autre candidat)."
+                    ]
                 ]
             ]
-        ]
-    ],
-    "skills" => [
-        "sidebar_title" => "Reconnaissance des compétences",
-        "category_anchor" => "skill-recognition",
-        "category_title" => "Comment fonctionne le modèle de reconnaissance des compétences du Nuage des talents?",
-        "sections" => [
-            "00" => [
-                "title" => "Dossiers de compétences dynamiques : Qu’est-ce que c’est?",
-                "copy" => [
-                    "00" => "C’est là que la magie se produit. Vous êtes plus qu’un CV et vos compétences sont plus que des titres de compétence officiels. Alors pourquoi ne pourriez-vous pas montrer à un employeur potentiel votre valeur réelle? Grâce aux dossiers de compétences dynamiques du Nuage des talents, vous pouvez être reconnu pour vos compétences, vos connaissances et vos capacités actuelles, peu importe comment, et où vous les avez obtenus. Cela signifie que vous pouvez revendiquer toutes vos expériences vécues comme moyen de démontrer votre expertise et ce dont vous êtes capable. Chaque fois que quelque chose est validé, elle fait partie de votre dossier de compétences dynamiques."
+        ],
+        "skills" => [
+            "sidebar_title" => "Reconnaissance des compétences",
+            "category_anchor" => "skill-recognition",
+            "category_title" => "Comment fonctionne le modèle de reconnaissance des compétences du Nuage des talents?",
+            "sections" => [
+                "00" => [
+                    "title" => "Dossiers de compétences dynamiques : Qu'est-ce que c'est?",
+                    "copy" => [
+                        "00" => "C'est là que la magie se produit. Vous êtes plus qu'un CV et vos compétences sont plus que des titres de compétence officiels. Alors pourquoi ne pourriez-vous pas montrer à un employeur potentiel votre valeur réelle? Grâce aux dossiers de compétences dynamiques du Nuage des talents, vous pouvez être reconnu pour vos compétences, vos connaissances et vos capacités actuelles, peu importe comment, et où vous les avez obtenus. Cela signifie que vous pouvez revendiquer toutes vos expériences vécues comme moyen de démontrer votre expertise et ce dont vous êtes capable. Chaque fois que quelque chose est validé, elle fait partie de votre dossier de compétences dynamiques."
+                    ]
+                ],
+                "01" => [
+                    "title" => "Voie de vie unique? Pas de problème!",
+                    "copy" => [
+                        "00" => "Vous avez un diplôme? Parfait. Vous n'en avez pas? Parfait. Nous nous soucions de ce dont vous êtes capable. Notre méthode dynamique de reconnaissance des compétences vous aide à nous le dire, quel que soit le chemin que vous avez choisi pour arriver ici. Bien que certains emplois exigent toujours un diplôme universitaire en particulier (selon les normes de classification du gouvernement). Le Nuage des talents aide les gestionnaires à élaborer une approche générale de reconnaissance des compétences et des équivalences. L'idée est de modifier la façon dont le gouvernement voit et valorise l'expérience afin qu'un plus large éventail de talents canadiens puisse s'appliquer au travail du gouvernement."
+                    ]
+                ],
+                "02" => [
+                    "title" => "Pourquoi vous en donnez la peine et qu'est-ce que vous en retirez?",
+                    "copy" => [
+                        "00" => "Reconnaissance : Vous devenez propriétaire de dossiers de compétences dynamiques reconnus qui tiennent compte de vos compétences et votre potentiel actuels.",
+                        "01" => "Instaurer la confiance : Les employeurs potentiels peuvent avoir confiance dans vos dossiers de compétences dynamiques parce qu'ils ont déjà été validés.",
+                        "02" => "Économies de temps : Vous gagnez beaucoup de temps en présentant une demande pour de nouvelles possibilités parce que vous pouvez réutiliser vos dossiers de compétences dynamiques."
+                    ]
+                ],
+                "03" => [
+                    "title" => "Comment cela fonctionne-t-il?",
+                    "copy" => [
+                        "00" => "C'est vraiment très simple!",
+                        "01" => [
+                            "type" => "ol",
+                            "items" => [
+                                "00" => "Réclamer : Vous nous parlez d'une compétence que vous possédez",
+                                "01" => "Recueillir : Vous nous envoyez un échantillon de travail qui démontre votre compétence",
+                                "02" => "Corroborer : Vous choisissez quelqu'un qui peut agir comme microréférence pour appuyer votre demande",
+                                "03" => "Réutiliser : La prochaine fois que vous postulez un emploi, vous pouvez tout simplement réutiliser ce dossier – il n'est pas nécessaire de réinventer la roue chaque fois!",
+                                "04" => "Évoluer : Vous pouvez ajouter des éléments à vos dossiers dynamiques de compétences afin qu'ils grandissent au fil du temps comme vous"
+                            ]
+                        ]
+                    ]
                 ]
-            ],
-            "01" => [
-                "title" => "Voie de vie unique? Pas de problème!",
-                "copy" => [
-                    "00" => "Vous avez un diplôme? Parfait. Vous n’en avez pas? Parfait. Nous nous soucions de ce dont vous êtes capable. Notre méthode dynamique de reconnaissance des compétences vous aide à nous le dire, quel que soit le chemin que vous avez choisi pour arriver ici. Bien que certains emplois exigent toujours un diplôme universitaire en particulier (selon les normes de classification du gouvernement). Le Nuage des talents aide les gestionnaires à élaborer une approche générale de reconnaissance des compétences et des équivalences. L’idée est de modifier la façon dont le gouvernement voit et valorise l’expérience afin qu’un plus large éventail de talents canadiens puisse s’appliquer au travail du gouvernement."
-                ]
-            ],
-            "02" => [
-                "title" => "Pourquoi vous en donnez la peine et qu’est-ce que vous en retirez?",
-                "copy" => [
-                    "00" => "Reconnaissance : Vous devenez propriétaire de dossiers de compétences dynamiques reconnus qui tiennent compte de vos compétences et votre potentiel actuels.",
-                    "01" => "Instaurer la confiance : Les employeurs potentiels peuvent avoir confiance dans vos dossiers de compétences dynamiques parce qu’ils ont déjà été validés.",
-                    "02" => "Économies de temps : Vous gagnez beaucoup de temps en présentant une demande pour de nouvelles possibilités parce que vous pouvez réutiliser vos dossiers de compétences dynamiques."
-                ]
-            ],
-            "03" => [
-                "title" => "Comment cela fonctionne-t-il?",
-                "copy" => [
-                    "00" => "C’est vraiment très simple!",
-                    "01" => [
-                        "type" => "ol",
-                        "items" => [
-                            "00" => "Réclamer : Vous nous parlez d’une compétence que vous possédez",
-                            "01" => "Recueillir : Vous nous envoyez un échantillon de travail qui démontre votre compétence",
-                            "02" => "Corroborer : Vous choisissez quelqu’un qui peut agir comme microréférence pour appuyer votre demande",
-                            "03" => "Réutiliser : La prochaine fois que vous postulez un emploi, vous pouvez tout simplement réutiliser ce dossier – il n’est pas nécessaire de réinventer la roue chaque fois!",
-                            "04" => "Évoluer : Vous pouvez ajouter des éléments à vos dossiers dynamiques de compétences afin qu’ils grandissent au fil du temps comme vous"
+            ]
+        ],
+        "levels" => [
+            "sidebar_title" => "Niveau de compétences",
+            "category_anchor" => "levels",
+            "category_title" => "Comment puis-je connaître le niveau de mes compétences?",
+            "sections" => [
+                "00" => [
+                    "title" => "Modèle de compétences du Nuage de talents",
+                    "copy" => [
+                        "00" => "Plutôt que de définir l'expérience en fonction du nombre d'années d'expérience, le Nuage de talents examine la capacité d'une personne à faire son travail dans diverses conditions de complexité et d'autonomie. Essentiellement, votre niveau est déterminé par la difficulté d'une tâche, qui est mesurée en fonction de l'aide ou de la supervision dont vous avez besoin pour l'accomplir. Cette approche reconnaît que les gens acquerront des compétences à des rythmes différents, et que leurs forces se manifesteront de différentes façons."
+                    ]
+                ],
+                "01" => [
+                    "title" => "Trouvez votre niveau: Compétences spécialisées",
+                    "copy" => [
+                        "00" => "Débutant :",
+                        "01" => [
+                            "type" => "ul",
+                            "items" => [
+                                "00" => "Vous êtes capable d'accomplir des tâches de base avec une supervision régulière et une orientation claire. Les tâches qui vous sont assignées sont claires et ne sont pas très complexes. Elles ont généralement une incidence locale.",
+                                "01" => "Au fur et à mesure que vous progressez dans cette catégorie, vous devriez être en mesure d'accomplir des tâches de complexité modérée avec une supervision régulière. Vous devriez aussi être en mesure d'accomplir des tâches de base avec peu ou pas de supervision.",
+                                "02" => "Ce niveau est habituellement associé aux tâches qui constituent le gros du travail pour les postes de niveau inférieur, comme les analystes ou les développeurs de niveau débutant."
+                            ]
+                        ],
+                        "02" => "Intermédiaire :",
+                        "03" => [
+                            "type" => "ul",
+                            "items" => [
+                                "00" => "Vous avez la capacité d'accomplir des tâches de complexité modérée ou d'incidence modérée avec supervision. C'est le superviseur qui détermine l'approche à préconiser pour effectuer les tâches, de même que la façon dont elles sont exécutées. Vous apportez des commentaires et des conseils. Vous êtes en mesure de faire progresser la tâche, même face à des obstacles et à des complications de petite à moyenne envergure.",
+                                "01" => "Au fur et à mesure que vous progressez dans cette catégorie, vous devriez être en mesure d'accomplir des tâches d'une complexité importante ou ayant une incidence plus grande avec une supervision régulière. Vous devriez également être en mesure d'accomplir des tâches d'une complexité ou d'une incidence modérée avec peu ou pas de supervision.",
+                                "02" => "Ce niveau est habituellement associé aux tâches qui constituent le gros du travail pour les postes de niveau intermédiaire, comme les analystes ou les développeurs."
+                            ]
+                        ],
+                        "04" => "Avancé :",
+                        "05" => [
+                            "type" => "ul",
+                            "items" => [
+                                "00" => "Vous avez la capacité d'accomplir des tâches d'une complexité ou d'une incidence importante avec supervision. Vous donnez des conseils et des commentaires au superviseur sur l'approche à employer pour effectuer les tâches et la façon dont elles sont exécutées. Vous êtes en mesure de faire progresser la tâche, même face à des obstacles et à des complications d'envergure moyenne à importante.",
+                                "01" => "Au fur et à mesure que vous progressez dans cette catégorie, vous êtes être en mesure d'accomplir des tâches d'une complexité importante ou ayant une incidence plus grande avec seulement de légers niveaux de supervision, en étant effectivement le responsable de l'initiative. Vous pouvez également jouer un rôle de formation d'autres personnes dans cet ensemble de compétences ou assumer un rôle de supervision léger pour les personnes aux niveaux inférieurs.",
+                                "02" => "Ce niveau est habituellement associé à des tâches qui constituent la majeure partie du travail pour des postes de niveau supérieur, comme les analystes principaux ou les développeurs principaux."
+                            ]
+                        ],
+                        "06" => "Responsable :",
+                        "07" => [
+                            "type" => "ul",
+                            "items" => [
+                                "00" => "Vous êtes en mesure d'accomplir des tâches d'une complexité ou d'une incidence importante, où vous prenez les décisions et répondez de vos décisions auprès de la haute direction de l'organisation. Vous présentez les tâches, l'approche et le plan de réalisation à la haute direction. Vous supervisez souvent d'autres personnes (personnes ou équipes) dans l'exécution de tâches très complexes ou ayant une incidence sur l'ensemble du système. Vous êtes en mesure de faire progresser ces tâches, même face à des obstacles et à des complications importants et imprévus.",
+                                "01" => "Au fur et à mesure que vous progressez dans cette catégorie, vous devriez être en mesure d'évaluer les autres à des niveaux plus subalternes, et de déterminer clairement la différence entre les tâches débutantes, intermédiaires et avancées. Vous devriez également être en mesure de pouvoir former des équipes, définir des orientations et assurer une supervision.",
+                                "02" => "Ce niveau est habituellement associé aux tâches qui constituent la majeure partie du travail pour les postes de direction et de direction."
+                            ]
+                        ]
+                    ]
+                ],
+                "02" => [
+                    "title" => "Trouvez votre niveau: Compétences non spécialisées",
+                    "copy" => [
+                        "00" => "Phase de développement précoce :",
+                        "01" => [
+                            "type" => "ul",
+                            "items" => [
+                                "00" => "Vous êtes en processus d'acquérir cette compétence ou cet attribut. Vous êtes capable de le démontrer dans des conditions favorables (peu de stress, difficulté minimale) et pouvez l'appliquer dans un contexte de travail de façon intermittente."
+                            ]
+                        ],
+                        "02" => "Modérément en évidence :",
+                        "03" => [
+                            "type" => "ul",
+                            "items" => [
+                                "00" => "Vous êtes capable de démontrer cette compétence ou cet attribut de façon constante en milieu de travail, y compris lorsque les conditions de difficulté ou le niveau de stress sont bas ou modérés.",
+                                "01" => "Vos pairs et vos superviseurs peuvent attester le fait que vous êtes capable de démontrer cette compétence ou cet attribut de façon régulière."
+                            ]
+                        ],
+                        "04" => "Fortement en évidence :",
+                        "05" => [
+                            "type" => "ul",
+                            "items" => [
+                                "00" => "Vous êtes capable de démontrer cette compétence ou cet attribut de façon constante en milieu de travail, y compris lorsque les conditions de difficulté ou le niveau de stress sont élevés.",
+                                "01" => "Vos pairs et vos superviseurs reconnaissent qu'il s'agit d'une force dont vous faites preuve en milieu de travail."
+                            ]
+                        ],
+                        "06" => "Démonstration à un niveau profond :",
+                        "07" => [
+                            "type" => "ul",
+                            "items" => [
+                                "00" => "Il s'agit d'une partie fondamentale de qui vous êtes. Vous démontrez cette compétence ou cet attribut de façon constante en milieu de travail, y compris lorsque les conditions de difficulté ou le niveau de stress sont extrêmes.",
+                                "01" => "Vos pairs et vos superviseurs reconnaissent qu'il s'agit d'une force importante dont vous faites preuve en milieu de travail, en donnant un exemple aux autres."
+                            ]
                         ]
                     ]
                 ]
             ]
         ]
-    ],
-    "levels" => [
-        "sidebar_title" => "Niveau de compétences",
-        "category_anchor" => "levels",
-        "category_title" => "Comment puis-je connaître le niveau de mes compétences?",
-        "sections" => [
-            "00" => [
-                "title" => "Modèle de compétences du Nuage de talents",
-                "copy" => [
-                    "00" => "Plutôt que de définir l’expérience en fonction du nombre d’années d’expérience, le Nuage de talents examine la capacité d’une personne à faire son travail dans diverses conditions de complexité et d’autonomie. Essentiellement, votre niveau est déterminé par la difficulté d’une tâche, qui est mesurée en fonction de l’aide ou de la supervision dont vous avez besoin pour l’accomplir. Cette approche reconnaît que les gens acquerront des compétences à des rythmes différents, et que leurs forces se manifesteront de différentes façons."
-                ]
-            ],
-            "01" => [
-                "title" => "Trouvez votre niveau: Compétences spécialisées",
-                "copy" => [
-                    "00" => "Débutant :",
-                    "01" => [
-                        "type" => "ul",
-                        "items" => [
-                            "00" => "Vous êtes capable d’accomplir des tâches de base avec une supervision régulière et une orientation claire. Les tâches qui vous sont assignées sont claires et ne sont pas très complexes. Elles ont généralement une incidence locale.",
-                            "01" => "Au fur et à mesure que vous progressez dans cette catégorie, vous devriez être en mesure d’accomplir des tâches de complexité modérée avec une supervision régulière. Vous devriez aussi être en mesure d’accomplir des tâches de base avec peu ou pas de supervision.",
-                            "02" => "Ce niveau est habituellement associé aux tâches qui constituent le gros du travail pour les postes de niveau inférieur, comme les analystes ou les développeurs de niveau débutant."
-                        ]
-                    ],
-                    "02" => "Intermédiaire :",
-                    "03" => [
-                        "type" => "ul",
-                        "items" => [
-                            "00" => "Vous avez la capacité d’accomplir des tâches de complexité modérée ou d’incidence modérée avec supervision. C’est le superviseur qui détermine l’approche à préconiser pour effectuer les tâches, de même que la façon dont elles sont exécutées. Vous apportez des commentaires et des conseils. Vous êtes en mesure de faire progresser la tâche, même face à des obstacles et à des complications de petite à moyenne envergure.",
-                            "01" => "Au fur et à mesure que vous progressez dans cette catégorie, vous devriez être en mesure d’accomplir des tâches d’une complexité importante ou ayant une incidence plus grande avec une supervision régulière. Vous devriez également être en mesure d’accomplir des tâches d’une complexité ou d’une incidence modérée avec peu ou pas de supervision.",
-                            "02" => "Ce niveau est habituellement associé aux tâches qui constituent le gros du travail pour les postes de niveau intermédiaire, comme les analystes ou les développeurs."
-                        ]
-                    ],
-                    "04" => "Avancé :",
-                    "05" => [
-                        "type" => "ul",
-                        "items" => [
-                            "00" => "Vous avez la capacité d’accomplir des tâches d’une complexité ou d’une incidence importante avec supervision. Vous donnez des conseils et des commentaires au superviseur sur l’approche à employer pour effectuer les tâches et la façon dont elles sont exécutées. Vous êtes en mesure de faire progresser la tâche, même face à des obstacles et à des complications d’envergure moyenne à importante.",
-                            "01" => "Au fur et à mesure que vous progressez dans cette catégorie, vous êtes être en mesure d’accomplir des tâches d’une complexité importante ou ayant une incidence plus grande avec seulement de légers niveaux de supervision, en étant effectivement le responsable de l’initiative. Vous pouvez également jouer un rôle de formation d’autres personnes dans cet ensemble de compétences ou assumer un rôle de supervision léger pour les personnes aux niveaux inférieurs.",
-                            "02" => "Ce niveau est habituellement associé à des tâches qui constituent la majeure partie du travail pour des postes de niveau supérieur, comme les analystes principaux ou les développeurs principaux."
-                        ]
-                    ],
-                    "06" => "Responsable :",
-                    "07" => [
-                        "type" => "ul",
-                        "items" => [
-                            "00" => "Vous êtes en mesure d’accomplir des tâches d’une complexité ou d’une incidence importante, où vous prenez les décisions et répondez de vos décisions auprès de la haute direction de l’organisation. Vous présentez les tâches, l’approche et le plan de réalisation à la haute direction. Vous supervisez souvent d’autres personnes (personnes ou équipes) dans l’exécution de tâches très complexes ou ayant une incidence sur l’ensemble du système. Vous êtes en mesure de faire progresser ces tâches, même face à des obstacles et à des complications importants et imprévus.",
-                            "01" => "Au fur et à mesure que vous progressez dans cette catégorie, vous devriez être en mesure d’évaluer les autres à des niveaux plus subalternes, et de déterminer clairement la différence entre les tâches débutantes, intermédiaires et avancées. Vous devriez également être en mesure de pouvoir former des équipes, définir des orientations et assurer une supervision.",
-                            "02" => "Ce niveau est habituellement associé aux tâches qui constituent la majeure partie du travail pour les postes de direction et de direction."
-                        ]
-                    ]
-                ]
-            ],
-            "02" => [
-                "title" => "Trouvez votre niveau: Compétences non spécialisées",
-                "copy" => [
-                    "00" => "Phase de développement précoce :",
-                    "01" => [
-                        "type" => "ul",
-                        "items" => [
-                            "00" => "Vous êtes en processus d’acquérir cette compétence ou cet attribut. Vous êtes capable de le démontrer dans des conditions favorables (peu de stress, difficulté minimale) et pouvez l’appliquer dans un contexte de travail de façon intermittente."
-                        ]
-                    ],
-                    "02" => "Modérément en évidence :",
-                    "03" => [
-                        "type" => "ul",
-                        "items" => [
-                            "00" => "Vous êtes capable de démontrer cette compétence ou cet attribut de façon constante en milieu de travail, y compris lorsque les conditions de difficulté ou le niveau de stress sont bas ou modérés.",
-                            "01" => "Vos pairs et vos superviseurs peuvent attester le fait que vous êtes capable de démontrer cette compétence ou cet attribut de façon régulière."
-                        ]
-                    ],
-                    "04" => "Fortement en évidence :",
-                    "05" => [
-                        "type" => "ul",
-                        "items" => [
-                            "00" => "Vous êtes capable de démontrer cette compétence ou cet attribut de façon constante en milieu de travail, y compris lorsque les conditions de difficulté ou le niveau de stress sont élevés.",
-                            "01" => "Vos pairs et vos superviseurs reconnaissent qu’il s’agit d’une force dont vous faites preuve en milieu de travail."
-                        ]
-                    ],
-                    "06" => "Démonstration à un niveau profond :",
-                    "07" => [
-                        "type" => "ul",
-                        "items" => [
-                            "00" => "Il s’agit d’une partie fondamentale de qui vous êtes. Vous démontrez cette compétence ou cet attribut de façon constante en milieu de travail, y compris lorsque les conditions de difficulté ou le niveau de stress sont extrêmes.",
-                            "01" => "Vos pairs et vos superviseurs reconnaissent qu’il s’agit d’une force importante dont vous faites preuve en milieu de travail, en donnant un exemple aux autres."
-                        ]
-                    ]
-                ]
-            ]
-        ]
-    ]
-];
+    ];
