@@ -74,13 +74,16 @@ return [
         'english_essential_context_02' => 'Vous pouvez présenter cette demande initiale dans la langue officielle de votre choix (français ou anglais).',
         'french_essential_context_01' => 'Ce poste exige une bonne maîtrise du français, tant à l’écrit qu’à l’oral. Dans le cadre de l’évaluation de vos compétences linguistiques, le gestionnaire d’embauche peut vous demander de suivre certaines étapes d’évaluation en français, comme des questions d’entrevue ou un examen.',
         'french_essential_context_02' => 'Vous pouvez présenter cette demande initiale dans la langue officielle de votre choix (français ou anglais).',
-        'bilingual_advanced_context_01' => 'Ce poste exige une maîtrise avancée du français et de l’anglais. Cela signifie que vous possédez de solides compétences en lecture, en écriture et en communication verbale dans les deux langues officielles, et que vous pouvez assumer les fonctions du poste en français ou en anglais. Dans le cadre de ce processus de sélection, vos compétences linguistiques seront évaluées par la %proficiency%.',
+        'english_or_french_context_01' => 'Pour ce poste, vous répondez aux exigences linguistiques si vous possédez de solides compétences en lecture, en rédaction et en communication verbale en français, en anglais ou dans les deux (bilingue).',
+        'english_or_french_context_02' => 'Vous pouvez présenter cette demande dans la langue officielle de votre choix (français ou anglais).',
+        'bilingual_intermediate_context_01' => 'Ce poste exige une connaissance pratique du français et de l’anglais. Cela signifie que vous pouvez assumer les fonctions du poste en français ou en anglais, et que vous possédez des compétences intermédiaires en lecture, en rédaction et en communication verbale dans les deux langues officielles. Dans le cadre de ce processus de sélection, vos compétences linguistiques seront évaluées par la %proficiency%.',
+        'bilingual_intermediate_context_02' => 'Vous pouvez suivre toutes les autres étapes de ce processus d’évaluation dans la langue officielle de votre choix, y compris la demande initiale, l’entrevue, l’examen et toute autre composante de l’évaluation.',
+        'bilingual_advanced_context_01' => 'Ce poste exige une connaissance avancé du français et de l’anglais. Cela signifie que vous pouvez assumer les fonctions du poste en français ou en anglais, et que vous possédez des solides compétences en lecture, en rédaction et en communication verbale dans les deux langues. Dans le cadre de ce processus de sélection, vos compétences linguistiques seront évaluées par la %proficiency%.',
         'bilingual_advanced_context_02' => 'Vous pouvez suivre toutes les autres étapes de ce processus d’évaluation dans la langue officielle de votre choix, y compris la demande initiale, l’entrevue, l’examen et toute autre composante de l’évaluation.',
         'proficiency_link' => 'https://www.canada.ca/fr/commission-fonction-publique/emplois/services/emplois-gc/renseignements-candidats/exigences-linguistiques.html',
         'proficiency_copy' => 'Commission de la fonction publique du Canada',
         'assessment_link' => 'https://www.canada.ca/fr/commission-fonction-publique/services/evaluation-langue-seconde/tests-autoevaluation.html',
         'assessment_copy' => 'tests d’autoévaluation',
-        'english_french_context' => 'TRANSLATION NEEDED: NEEDS COPY'
     ],
     'culture' => [
         'sidebar_title' => 'Voir la culture à cet emploi.',
